pyproject.toml
[project]
name = "gitlite-backkend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio==4.11.0",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "click==8.3.0",
    "colorama==0.4.6",
    "cryptography==46.0.3",
    "deprecation==2.1.0",
    "email-validator>=2.3.0",
    "fastapi==0.119.1",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "hyperframe==6.1.0",
    "multidict==6.7.0",
    "packaging==25.0",
    "postgrest==2.22.0",
    "propcache==0.4.1",
    "pycparser==2.23",
    "pydantic-settings==2.11.0",
    "pydantic[email]>=2.12.3",
    "pyjwt==2.10.1",
    "python-dotenv==1.1.1",
    "realtime==2.22.0",
    "sniffio==1.3.1",
    "starlette==0.48.0",
    "storage3==2.22.0",
    "strenum==0.4.15",
    "supabase==2.22.0",
    "supabase-auth==2.22.0",
    "supabase-functions==2.22.0",
    "uvicorn==0.38.0",
    "websockets==15.0.1",
    "yarl==1.22.0",
]
